refactor: move benchmarks into separate module
