+++ .github/dependabot.yml
@@ -6,6 +6,12 @@ updates:
       interval: daily
     open-pull-requests-limit: 10
 
+  - package-ecosystem: gomod
+    directory: benchmarks
+    schedule:
+      interval: daily
+    open-pull-requests-limit: 10
+
   - package-ecosystem: github-actions
     directory: /
     schedule:
