+++ benchmarks/go.mod
@@ -1,0 +1,40 @@
+// benchmarks in separate module to avoid needless dependencies
+
+module benchmarks
+
+go 1.18
+
+replace github.com/abemedia/go-don => ../
+
+require (
+	github.com/abemedia/go-don v0.0.0-00010101000000-000000000000
+	github.com/gin-gonic/gin v1.8.1
+	github.com/valyala/fasthttp v1.41.0
+)
+
+require (
+	github.com/abemedia/fasthttpfs v0.0.0-20220321013016-a7e6ad30856d // indirect
+	github.com/abemedia/httprouter v0.0.0-20220322190908-e57cbe4767db // indirect
+	github.com/andybalholm/brotli v1.0.4 // indirect
+	github.com/gin-contrib/sse v0.1.0 // indirect
+	github.com/go-playground/locales v0.14.0 // indirect
+	github.com/go-playground/universal-translator v0.18.0 // indirect
+	github.com/go-playground/validator/v10 v10.10.0 // indirect
+	github.com/goccy/go-json v0.9.11 // indirect
+	github.com/json-iterator/go v1.1.12 // indirect
+	github.com/klauspost/compress v1.15.9 // indirect
+	github.com/leodido/go-urn v1.2.1 // indirect
+	github.com/mattn/go-isatty v0.0.14 // indirect
+	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
+	github.com/modern-go/reflect2 v1.0.2 // indirect
+	github.com/pelletier/go-toml/v2 v2.0.1 // indirect
+	github.com/ugorji/go/codec v1.2.7 // indirect
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	golang.org/x/crypto v0.0.0-20220214200702-86341886e292 // indirect
+	golang.org/x/net v0.0.0-20220906165146-f3363e06e74c // indirect
+	golang.org/x/sys v0.0.0-20220728004956-3c1f35247d10 // indirect
+	golang.org/x/text v0.3.7 // indirect
+	google.golang.org/protobuf v1.28.0 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+)
