+++ errors.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/goccy/go-json"
 	"github.com/valyala/fasthttp"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
 func E(err error) fasthttp.RequestHandler {
