+++ go.mod
@@ -4,36 +4,18 @@ go 1.18
 
 require (
 	github.com/abemedia/httprouter v0.0.0-20220322190908-e57cbe4767db
-	github.com/gin-gonic/gin v1.8.2
 	github.com/goccy/go-json v0.10.0
-	github.com/goccy/go-reflect v1.2.0
 	github.com/google/go-cmp v0.5.9
 	github.com/gorilla/schema v1.2.0
 	github.com/valyala/fasthttp v1.44.0
 	github.com/vmihailenco/msgpack/v5 v5.3.5
-	gopkg.in/yaml.v2 v2.4.0
+	gopkg.in/yaml.v3 v3.0.1
 )
 
 require (
 	github.com/abemedia/fasthttpfs v0.0.0-20220321013016-a7e6ad30856d // indirect
 	github.com/andybalholm/brotli v1.0.4 // indirect
-	github.com/gin-contrib/sse v0.1.0 // indirect
-	github.com/go-playground/locales v0.14.0 // indirect
-	github.com/go-playground/universal-translator v0.18.0 // indirect
-	github.com/go-playground/validator/v10 v10.11.1 // indirect
-	github.com/json-iterator/go v1.1.12 // indirect
 	github.com/klauspost/compress v1.15.9 // indirect
-	github.com/leodido/go-urn v1.2.1 // indirect
-	github.com/mattn/go-isatty v0.0.16 // indirect
-	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
-	github.com/modern-go/reflect2 v1.0.2 // indirect
-	github.com/pelletier/go-toml/v2 v2.0.6 // indirect
-	github.com/ugorji/go/codec v1.2.7 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/vmihailenco/tagparser/v2 v2.0.0 // indirect
-	golang.org/x/crypto v0.0.0-20220214200702-86341886e292 // indirect
-	golang.org/x/net v0.4.0 // indirect
-	golang.org/x/sys v0.3.0 // indirect
-	golang.org/x/text v0.5.0 // indirect
-	google.golang.org/protobuf v1.28.1 // indirect
 )
