+++ handler_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/abemedia/go-don"
 	_ "github.com/abemedia/go-don/encoding/json"
 	_ "github.com/abemedia/go-don/encoding/text"
-	"github.com/abemedia/go-don/internal/httptest"
+	"github.com/abemedia/go-don/pkg/httptest"
 	"github.com/abemedia/httprouter"
 	"github.com/google/go-cmp/cmp"
 	"github.com/valyala/fasthttp"
