+++ internal/byteconv/byteconv.go
@@ -1,4 +1,4 @@
-package internal
+package byteconv
 
 import (
 	"reflect"
